requirements.txt
# Core ML - Python 3.13.5 requires PyTorch nightly
# pip install --pre torch torchvision torchaudio 

# Diffusion and AI
diffusers
transformers
accelerate
datasets
peft
tokenizers

# Optimization
onnx
onnxruntime-gpu
# Note: TensorRT requires manual installation for Windows

# Core Libraries
numpy
pandas
matplotlib
seaborn
pillow
opencv-python
scipy

# Utilities
tqdm
pyyaml
wandb
huggingface-hub
safetensors
filelock
packaging
regex
requests

# Testing
pytest
pytest-cov
pytest-benchmark

# Development
black
flake8
mypy
pre-commit

# Windows specific
pywin32
colorama

# Optional accelerations
bitsandbytes

# pip install torch-tensorrt --no-deps Install Torch-TensorRT separately 